Struct emulator::cpu::Cpu

pub struct Cpu {
    // some fields omitted
}

Cpu of the 2i.

Represents the 8 bit cpu of the 2i with 8 registers that are 8 bit wide and the three status registers (carry, negative, zero).

Examples:

use emulator::{Cpu, Instruction, Ram};

let mut cpu = Cpu::new();
let mut ram = Ram::new();

// R0 = 6
let inst = Instruction::new(0b00_00000_00_000_0110_01_01_0001_0).unwrap();

let _ = cpu.execute_instruction(inst, &mut ram);
assert_eq!(6, cpu.inspect_registers()[0]);

Methods

impl Cpu

fn new() -> Cpu

Create a new cpu with all registers and flags set to zero.

fn execute_instruction<B: Bus>(&mut self, inst: Instruction, bus: &mut B) -> Result<(usize, Flags)>

Execute the given instruction on the cpu using the given, bus, input and output. Returns the address of the next instruction and the alu flags.

fn inspect_registers(&mut self) -> &mut [u8; 8]

Direct access to the registers.

fn inspect_flags(&mut self) -> &mut Flags

Direct access to the flag register.