Struct emulator::instruction::Instruction

pub struct Instruction {
    // some fields omitted
}

Instruction of the 2i.

Represents a 25 bit wide instruction that the 2i uses. Provides some conveniance methods to extract all different parts.

Methods

impl Instruction

fn new(instruction: u32) -> Result<Instruction>

Create a new Instruction from a u32. Fails if more than 25 bits are used.

fn new_from_string(string: &str) -> Result<Instruction>

Creat a new Instruction from a binary string (consisting only of ones and zeroes). Failes if more than 25 bits (characters) are used.

fn get_instruction(&self) -> u32

Get the instruction as a 25 bit integer (the first 7 most significant bits of the u32 are always zero)

fn should_store_flags(&self) -> bool

MCHFLG

fn get_alu_instruction(&self) -> u8

MALUS0-3 (4 bit)

fn is_alu_input_b_const(&self) -> bool

MALUIB

fn is_alu_input_a_bus(&self) -> bool

MALUIA

fn should_write_register(&self) -> bool

MRGWE

fn should_write_register_b(&self) -> bool

MRGWS

fn get_register_address_b(&self) -> usize

MRGAB0-2 (3 bit)

fn get_constant_input(&self) -> u8

MRGAB0-3 (4 bit)

fn get_register_address_a(&self) -> usize

MRGAA0-2 (3 bit)

fn is_bus_enabled(&self) -> bool

BUSEN

fn is_bus_writable(&self) -> bool

BUSWR

fn get_next_instruction_address(&self) -> u8

NA0-4 (5 bit)

fn get_address_control(&self) -> u8

MAC0-1 (2 bit)

fn to_text_paraphrase(&self, next_address: Option<usize>) -> String

Create a textual representation of the instruction

Optionally, the address of the next instruction after this one can be passed to simplify the resulting string in many cases.

Examples

use emulator::Instruction;

let inst = Instruction::new(0b00_00000_00_000_0110_01_01_0100_0).unwrap();
assert_eq!("R0 = R0 + 6", &inst.to_text_paraphrase(Some(0)));