Struct emulator::bus::Ram

pub struct Ram<'a> { /* fields omitted */ }

Ram of the 2i.

Represents the 8 bit ram of the 2i.

Methods

impl<'a> Ram<'a>

fn new() -> Ram<'a>

Create a new ram with all addresses initialised to zero.

fn inspect(&self) -> &RefCell<[u8; 256]>

Direct access to the ram wrapped in a RefCell.

fn add_overlay(&mut self,
first_address: u8,
last_address: u8,
overlay_bus: &'a Bus)

Add a bus as an overlay to the ram.

When a read or write lies in the given (inclusive) range, the request is forwarded to the given bus. All overlays are checked in the order they were added.